+++ CHANGELOG.md
@@ -38,7 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add job scheduler basic logic [(#671)](https://github.com/wazuh/wazuh-indexer-plugins/pull/671)
 - Init content from snapshot [(#670)](https://github.com/wazuh/wazuh-indexer-plugins/pull/670)
 - Add scheduled content update [(#682)](https://github.com/wazuh/wazuh-indexer-plugins/pull/682)
-- Add space-wide content management [(#684)](https://github.com/wazuh/wazuh-indexer-plugins/pull/684)
+- Implement space-wide content management [(#684)](https://github.com/wazuh/wazuh-indexer-plugins/pull/684)
+- Implement hash of hashes calculation [(#686)](https://github.com/wazuh/wazuh-indexer-plugins/pull/686)
 
 ### Dependencies
 - Upgrade to Gradle 8.14.3 [(#649)](https://github.com/wazuh/wazuh-indexer-plugins/pull/649)